Fix upper group limit in GroupedTag (#2996)
* Fix upper group limit to allow for 2^31 components.

Previously the formula for increasing the size of Uint32Arrays
in the GroupedTag was a little aggressive and would cut of very
short of 2^31 groups. Instead it would cut off between 2^12 and
2^13 components, which isn't quite a lot.

This new upper boundary is the absolute limit of how many slots
a Uint32Array is defined to be able to hold. After this limit we'll
see an error.

Instead of letting this error leak out we now throw a more clear
error that hints at this limitation.

* Add test for upper group limit

* Add additional check for GroupIDAllocator

We can make it easier to track the offending
dynamic components in development by adding another
SMI overflow check to the GroupIDAllocator

* Replace TypeError in dom utils with throwStyledError

* changelog entry

Co-authored-by: Evan Jacobs <probablyup@gmail.com>