refactor(types): fix attrs hole-punching
There is an edge case documented in attrs.test.tsx where a base
component may have a type that marks a particular prop as required.

Using attrs you can provide a value for this required field, in that
event we need to mark the field value in the base type as optional
since its known that a value was provided.